+++ apps/meteor/server/methods/addAllUserToRoom.ts
@@ -1,7 +1,7 @@
 import { Message } from '@rocket.chat/core-services';
 import type { IRoom } from '@rocket.chat/core-typings';
+import type { ServerMethods } from '@rocket.chat/ddp-client';
 import { Subscriptions, Rooms, Users } from '@rocket.chat/models';
-import type { ServerMethods } from '@rocket.chat/ui-contexts';
 import { check } from 'meteor/check';
 import { Meteor } from 'meteor/meteor';
 
@@ -11,7 +11,7 @@ import { getDefaultSubscriptionPref } from '../../app/utils/lib/getDefaultSubscr
 import { callbacks } from '../../lib/callbacks';
 import { getSubscriptionAutotranslateDefaultConfig } from '../lib/getSubscriptionAutotranslateDefaultConfig';
 
-declare module '@rocket.chat/ui-contexts' {
+declare module '@rocket.chat/ddp-client' {
 	// eslint-disable-next-line @typescript-eslint/naming-convention
 	interface ServerMethods {
 		addAllUserToRoom(rid: IRoom['_id'], activeUsersOnly?: boolean): Promise<true>;