+++ packages/rest-typings/src/v1/omnichannel.ts
@@ -2595,6 +2595,7 @@ export type GETLivechatRoomsParams = PaginatedRequest<{
 	departmentId?: string;
 	open?: string | boolean;
 	onhold?: string | boolean;
+	queued?: string | boolean;
 	tags?: string[];
 }>;
 
@@ -2655,6 +2656,12 @@ const GETLivechatRoomsParamsSchema = {
 			],
 		},
 		onhold: {
+			anyOf: [
+				{ type: 'string', nullable: true },
+				{ type: 'boolean', nullable: true },
+			],
+		},
+		queued: {
 			anyOf: [
 				{ type: 'string', nullable: true },
 				{ type: 'boolean', nullable: true },