+++ ee/packages/license/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rocket.chat/license
 
+## 0.2.1
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@6.10.1
+  </details>
+
 ## 0.2.0
 
 ### Minor Changes