+++ packages/ui-contexts/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @rocket.chat/ui-contexts
 
+## 8.0.1
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@6.10.1
+  - @rocket.chat/rest-typings@6.10.1
+  - @rocket.chat/ddp-client@0.3.1
+  </details>
+
 ## 8.0.0
 
 ### Minor Changes